変更内容: - `api/src/utils/validation.ts` から `import.meta.vitest` ブロックを削除し、本体コードのみの状態に整理。 - 同じテストケースを `api/src/utils/validation.test.ts` に移動し、各バリデーション関数の正常系・異常系・メッセージ生成を網羅。
実行したコマンド:
- `pnpm -C api run format`
- `pnpm -C api run test`（テストはすべて pass。サービス層のエラーパスを検証するテストで stderr に例外ログが出ていますが、テスト意図どおりの挙動です）

必要があれば、この状態でコミットしてください。
